backend/requirements.txt
fastapi
uvicorn
anthropic
openai
pydantic
python-dotenv
pandas